added tests, updated docs
