+++ swxsoc/util/tests/test_util.py
@@ -9,8 +9,10 @@ from moto import mock_aws
 from moto.core import DEFAULT_ACCOUNT_ID as ACCOUNT_ID
 from moto.timestreamwrite.models import timestreamwrite_backends
 
-
+from astropy import units as u
+from astropy.timeseries import TimeSeries
 from astropy.time import Time
+
 import swxsoc
 from swxsoc.util import util
 
@@ -422,132 +424,3 @@ def test_create_configdir_configured(instrument, time, level, version, result):
     del os.environ["SWXSOC_CONFIGDIR"]
     swxsoc._reconfigure()
 # fmt: on
-
-
-@mock_aws
-def test_record_dimension_timestream():
-    # Create a timestream database with boto3
-    client = boto3.client("timestream-write", region_name="us-east-1")
-    database_name = "dev-swxsoc_sdc_aws_logs"
-    table_name = "dev-swxsoc_measures_table"
-    client.create_database(DatabaseName=database_name)
-
-    # Create a table in the database
-    client.create_table(
-        DatabaseName=database_name,
-        TableName=table_name,
-        RetentionProperties={
-            "MemoryStoreRetentionPeriodInHours": 24,
-            "MagneticStoreRetentionPeriodInDays": 7,
-        },
-    )
-
-    instrument_name = "eea"
-    dimensions = [
-        {"Name": "Location", "Value": "Mars"},
-        {"Name": "SensorType", "Value": "Temperature"},
-        {"Name": "Unit", "Value": "Celsius"},
-    ]
-    measure_name = "temperature_reading"
-    measure_value = 25.2
-    measure_value_type = "DOUBLE"
-
-    # Call the function to record dimensions in Timestream
-    util.record_dimension_timestream(
-        dimensions=dimensions,
-        instrument_name=instrument_name,
-        measure_name=measure_name,
-        measure_value=measure_value,
-        measure_value_type=measure_value_type,
-    )
-
-    instrument_name = "spani"
-    dimensions = [
-        {"Name": "Location", "Value": "Earth"},
-        {"Name": "Activity", "Value": "Running"},
-        {"Name": "Unit", "Value": "Seconds"},
-    ]
-
-    # Timestamp in milliseconds
-    timestamp = "1628460000000"
-
-    # Call the function to record the time as a measure in Timestream
-    util.record_dimension_timestream(
-        dimensions=dimensions,
-        instrument_name=instrument_name,
-        timestamp=timestamp,
-    )
-
-    backend = timestreamwrite_backends[ACCOUNT_ID]["us-east-1"]
-    records = backend.databases[database_name].tables[table_name].records
-
-    assert len(records) == 2
-
-
-@mock_aws
-def test_invalid_record_dimension_timestream():
-    # Create a timestream database with boto3
-    client = boto3.client("timestream-write", region_name="us-east-1")
-    database_name = "dev-swxsoc_sdc_aws_logs"
-    table_name = "dev-swxsoc_measures_table"
-    client.create_database(DatabaseName=database_name)
-
-    # Create a table in the database
-    client.create_table(
-        DatabaseName=database_name,
-        TableName=table_name,
-        RetentionProperties={
-            "MemoryStoreRetentionPeriodInHours": 24,
-            "MagneticStoreRetentionPeriodInDays": 7,
-        },
-    )
-
-    instrument_name = "invalid_instrument"
-    dimensions = [
-        {"Name": "Location", "Value": "Mars"},
-        {"Name": "SensorType", "Value": "Temperature"},
-        {"Name": "Unit", "Value": "Celsius"},
-    ]
-
-    # Call the function to record dimensions in Timestream
-    util.record_dimension_timestream(
-        dimensions=dimensions,
-    )
-
-    backend = timestreamwrite_backends[ACCOUNT_ID]["us-east-1"]
-    record = backend.databases[database_name].tables[table_name].records[0]
-
-    record_dimensions = record["Dimensions"]
-
-    # Check if InstrumentName is in the record dimensions
-    assert "InstrumentName" not in record_dimensions
-
-
-@mock_aws
-def test_invalid_instrument_record_dimension_timestream():
-    # Create a timestream database with boto3
-    client = boto3.client("timestream-write", region_name="us-east-1")
-    database_name = "dev-swxsoc_sdc_aws_logs"
-    table_name = "dev-swxsoc_measures_table"
-    client.create_database(DatabaseName=database_name)
-
-    # Create a table in the database
-    client.create_table(
-        DatabaseName=database_name,
-        TableName=table_name,
-        RetentionProperties={
-            "MemoryStoreRetentionPeriodInHours": 24,
-            "MagneticStoreRetentionPeriodInDays": 7,
-        },
-    )
-
-    dimensions = "invalid"
-
-    # Call the function to record dimensions in Timestream
-    util.record_dimension_timestream(dimensions=dimensions)
-
-    backend = timestreamwrite_backends[ACCOUNT_ID]["us-east-1"]
-    records = backend.databases[database_name].tables[table_name].records
-
-    # Assert that no records were created
-    assert len(records) == 0
